backend/requirements.txt
uvicorn~=0.40.0
fastapi~=0.128.0
sqlalchemy~=2.0.44
alembic~=1.17.1
asyncpg~=0.30.0
psycopg2-binary~=2.9.10
pydantic-settings~=2.11.0
